Refactor NeoFaker.Time
