+++ lib/neo_faker/time/between.ex
@@ -1,20 +1,8 @@
-defmodule NeoFaker.Time.Utils do
+defmodule NeoFaker.Time.Between do
   @moduledoc false
 
   @type time_format :: :sigil | :iso8601
   @type time_unit :: :hour | :minute | :second
-
-  @doc """
-  Generate a random time with sepecified day range.
-  """
-  @spec random_add_time(Range.t(), time_unit(), time_format()) :: Time.t() | String.t()
-  def random_add_time(range, unit, :sigil) do
-    NaiveDateTime.local_now() |> Time.add(Enum.random(range), unit) |> Time.truncate(:second)
-  end
-
-  def random_add_time(range, unit, :iso8601) do
-    range |> random_add_time(unit, :sigil) |> Time.to_iso8601()
-  end
 
   @doc """
   Generate a random time with a specific between two times.
@@ -27,11 +15,9 @@ defmodule NeoFaker.Time.Utils do
     {start_seconds, finish_seconds} =
       Enum.min_max([time_to_seconds_start, time_to_seconds_finish])
 
-    Time.add(
-      start,
-      Enum.random(start_seconds..finish_seconds) - start_seconds,
-      :second
-    )
+    amount_to_add = Enum.random(start_seconds..finish_seconds) - start_seconds
+
+    Time.add(start, amount_to_add, :second)
   end
 
   def random_between_time(start, finish, :iso8601) do
